+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - The `layer.popup` gets more types and uses `IPopupParams` which will be applied to all popups.
 
 * **@dlr-eoc/map-ol:**
+  - Popup events can be observed by `MapOlService.popupEvents: Subject<IPopupEvent>`. For this add `asObservable: true` to a popup object, you will then be able to subscribe to the event instead of having a popup container being added to the map.
   - A new Interface `IPopupEvent` is exposed.
   - Add groupID to olLayers (layer.groupID) to check if a layer was in an olLayerGroup.